requirements.txt
Django==3.2.10
environs==5.2.1
dj-database-url==0.5.0
django-phonenumber-field==8.0.0
phonenumber==8.13.51